initial commit moving this to a clean repo and organizing the project a bit. I wanna do some refactoring and renaming then try to publish this on github.
The idea is to use the latest gen ecmascript with native modules and make a simple console app framework for the browser with minimal dependencies.

this can for example be useful as a teaching tool since console apps are rather less complicated to write and the setup can be a lot easier at least if your used to more web relates stuff.

I also have some ideas about using this for a text based rogue like game.